DE-6423: Improve types, add lint
`@castlabs/prestoplay` is untyped, this leads to large parts
of the codebase being untyped as a result.
Fix it by adding basic typings for `@castlabs/prestoplay`.
Fix several problems discovered thanks to the new types.

Add ESLint with rules for TS, React and React hooks.
Fix several problems discovered thanks to the new lint.

*Note that there are still some @ts-ignores and not all lint problems
are fixed. Those are cases that are non-trivial to fix and should be
fixed carefully in a dedicated PR later.

App Rollup config `scripts/rollup.config.app.mjs` is confusing
in that it duplicates configuration which is already defined
in `tscfonfig.json`.
Simplify this by removing the duplicated config.

In a few places the code manipulates the DOM directly, this is
brittle, because React engine does not know about this and so
React can unintentionally break that part of the DOM.
Fix some of those places by rewriting it to preform that
DOM manipulation via React. (e.g. using React Helmet)

Extract `TrackLabeller` into a dedicated file to simplify the codebase.

Simplify the logic of the `Thumbnail` component and handle corner cases.

Make sure CI tests the code.