DE-6498: Add Storybook (#12)
Add Storybook as a framework to develop and document UI.
So far this is groundwork, once the Storybook docs are finished,
`./app` folder can be deleted.

* FIXES

Several small fixes to some components such as `Duration` and
`CurrentTime`, fixes to the UI and some improvements to the API.

* FEATURES

Create `PlayPauseIndicator`.
Add CSS classes `.pp-ui-layers` and `.pp-ui-layer` intended
to stack layers of UI on top of each other (e.g. player and
its overlay).

* TOOLING

In Jest remove `transformIgnorePatterns` of `@castlabs/prestoplay`
because it is not being transformed anyway, because it is getting
mocked instead.