DE-6746: Add Release workflow
Add a GitHub workflow which publishes the package to NPM
and API docs to `https://players.castlabs.com/react-dom/<version>/docs/`
when a new Release is manually created on GitHub.
The version is taken from the tag created with the Release on GitHub.

Beta versions are also supported.

Also clean up the readme and the docs homepage.