DE-6806: Use ESM entrypoint only
Currently we have a couple of entrypoints (browser, module, main) and
which entrypoint is used that depends solely on the bundler.

With the current package.json we have a serious problem for Webpack,
because it will by default select our "browser" entrypoint (CJS) and
for our dependency `@react-hook/resize-observer` it will select the
"module" entrypoint (ESM).

This mixing of CJS with ESM typically does not work correctly, at least
not when using default imports/exports. This is the case for both
Webpack and Rollup, it doesn't work correctly there.

In our case, that's exactly what is happening, we have a default
import from the `@react-hook/resize-observer` library:

```
import useResizeObserver from '@react-hook/resize-observer'
```

To avoid running into this problem, change the entrypoint to always
point to the ESM build.

This should be a safe solution assuming our dependencies also correctly
provide an ESM variant and assuming that the bundler supports ESM, which
I think is a safe assumption.